Refactor mysql dependencies out of util
