+++ python/hsfs/util.py
@@ -15,10 +15,8 @@
 #
 from __future__ import annotations
 
-import asyncio
 import itertools
 import json
-import logging
 import re
 import sys
 import threading
@@ -27,21 +25,15 @@ from datetime import date, datetime, timezone
 from typing import Any, Callable, Dict, List, Literal, Optional, Set, Tuple, Union
 from urllib.parse import urljoin, urlparse
 
-import numpy as np
 import pandas as pd
-from aiomysql.sa import create_engine as async_create_engine
 from hsfs import client, feature, feature_group, serving_key
 from hsfs.client import exceptions
 from hsfs.client.exceptions import FeatureStoreException
 from hsfs.constructor import serving_prepared_statement
 from hsfs.core import feature_group_api, variable_api
-from sqlalchemy import create_engine
-from sqlalchemy.engine.url import make_url
 
 
 FEATURE_STORE_NAME_SUFFIX = "_featurestore"
-
-_logger = logging.getLogger(__name__)
 
 
 class FeatureStoreEncoder(json.JSONEncoder):
@@ -128,50 +120,6 @@ def strip_feature_store_suffix(name: str) -> str:
         return name
 
 
-def create_mysql_engine(
-    online_conn: Any, external: bool, options: Optional[Dict[str, Any]] = None
-) -> Any:
-    online_options = online_conn.spark_options()
-    # Here we are replacing the first part of the string returned by Hopsworks,
-    # jdbc:mysql:// with the sqlalchemy one + username and password
-    # useSSL and allowPublicKeyRetrieval are not valid properties for the pymysql driver
-    # to use SSL we'll have to something like this:
-    # ssl_args = {'ssl_ca': ca_path}
-    # engine = create_engine("mysql+pymysql://<user>:<pass>@<addr>/<schema>", connect_args=ssl_args)
-    if external:
-        # This only works with external clients.
-        # Hopsworks clients should use the storage connector
-        host = get_host_name()
-        online_options["url"] = re.sub(
-            "/[0-9.]+:",
-            "/{}:".format(host),
-            online_options["url"],
-        )
-
-    sql_alchemy_conn_str = (
-        online_options["url"]
-        .replace(
-            "jdbc:mysql://",
-            "mysql+pymysql://"
-            + online_options["user"]
-            + ":"
-            + online_options["password"]
-            + "@",
-        )
-        .replace("useSSL=false&", "")
-        .replace("?allowPublicKeyRetrieval=true", "")
-    )
-    if options is not None and not isinstance(options, dict):
-        raise TypeError("`options` should be a `dict` type.")
-    if not options:
-        options = {"pool_recycle": 3600}
-    elif "pool_recycle" not in options:
-        options["pool_recycle"] = 3600
-    # default connection pool size kept by engine is 5
-    sql_alchemy_engine = create_engine(sql_alchemy_conn_str, **options)
-    return sql_alchemy_engine
-
-
 def get_host_name() -> str:
     host = variable_api.VariableApi().get_loadbalancer_external_domain()
     if host == "":
@@ -186,48 +134,6 @@ def get_dataset_type(path: str) -> Literal["HIVEDB", "DATASET"]:
         return "HIVEDB"
     else:
         return "DATASET"
-
-
-async def create_async_engine(
-    online_conn: Any,
-    external: bool,
-    default_min_size: int,
-    options: Optional[Dict[str, Any]] = None,
-    hostname: Optional[str] = None,
-) -> Any:
-    try:
-        loop = asyncio.get_running_loop()
-    except RuntimeError as er:
-        raise RuntimeError(
-            "Event loop is not running. Please invoke this co-routine from a running loop or provide an event loop."
-        ) from er
-
-    online_options = online_conn.spark_options()
-    # read the keys user, password from online_conn as use them while creating the connection pool
-    url = make_url(online_options["url"].replace("jdbc:", ""))
-    if hostname is None:
-        if external:
-            hostname = get_host_name()
-        else:
-            hostname = url.host
-
-    # create a aiomysql connection pool
-    pool = await async_create_engine(
-        host=hostname,
-        port=3306,
-        user=online_options["user"],
-        password=online_options["password"],
-        db=url.database,
-        loop=loop,
-        minsize=(
-            options.get("minsize", default_min_size) if options else default_min_size
-        ),
-        maxsize=(
-            options.get("maxsize", default_min_size) if options else default_min_size
-        ),
-        pool_recycle=(options.get("pool_recycle", -1) if options else -1),
-    )
-    return pool
 
 
 def check_timestamp_format_from_date_string(input_date: str) -> Tuple[str, str]:
@@ -282,13 +188,15 @@ def get_timestamp_from_date_string(input_date: str) -> int:
 
 
 def get_hudi_datestr_from_timestamp(timestamp: int) -> str:
-    return datetime.utcfromtimestamp(timestamp / 1000).strftime("%Y%m%d%H%M%S%f")[:-3]
+    return datetime.fromtimestamp(timestamp / 1000, timezone.utc).strftime(
+        "%Y%m%d%H%M%S%f"
+    )[:-3]
 
 
 def get_delta_datestr_from_timestamp(timestamp: int) -> str:
-    return datetime.utcfromtimestamp(timestamp / 1000).strftime("%Y-%m-%d %H:%M:%S.%f")[
-        :-3
-    ]
+    return datetime.fromtimestamp(timestamp / 1000, timezone.utc).strftime(
+        "%Y-%m-%d %H:%M:%S.%f"
+    )[:-3]
 
 
 def convert_event_time_to_timestamp(
@@ -548,24 +456,6 @@ def is_runtime_notebook():
         return True
     else:
         return False
-
-
-class NpDatetimeEncoder(json.JSONEncoder):
-    def default(self, obj):
-        dtypes = (np.datetime64, np.complexfloating)
-        if isinstance(obj, (datetime, date)):
-            return convert_event_time_to_timestamp(obj)
-        elif isinstance(obj, dtypes):
-            return str(obj)
-        elif isinstance(obj, np.integer):
-            return int(obj)
-        elif isinstance(obj, np.floating):
-            return float(obj)
-        elif isinstance(obj, np.ndarray):
-            if any([np.issubdtype(obj.dtype, i) for i in dtypes]):
-                return obj.astype(str).tolist()
-            return obj.tolist()
-        return super(NpDatetimeEncoder, self).default(obj)
 
 
 class VersionWarning(Warning):
